+++ src/ffrecv.h
@@ -9,7 +9,7 @@
 typedef struct ffrecv{    
     ffpeer_t peer;
     ffbuffer_t * buffer;
-    uint16_t tag;
+    uint8_t tag;
     uint8_t flags;
     uint8_t tag_type;
